pyproject.toml
[tool.poetry]
name = "stackoverflow-obj-det-api"
version = "0.1.0"
description = "Stack Overflow Object Detection API"
authors = ["BlaiseMuhirwa <blaisemuhirwa3@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
torch = "2.2.1"
black = "^24.4.2"
torchvision = "0.17.1"
pydantic = "^2.8.2"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
